backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-pptx==0.6.23
scikit-learn==1.5.1
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
spacy==3.7.5
vaderSentiment==3.3.2
joblib==1.4.2
orjson==3.10.7
starlette==0.38.5
google-generativeai==0.7.2
python-dotenv==1.0.1
